requirements.txt
pip==19.3.1
pandas==0.25.1
pygal==2.4.0
django_pandas==0.6.1
django_heroku==0.3.1
Django==2.2.4
argon2-cffi==19.1.0
bcrypt==3.1.7
cffi==1.12.3
gunicorn==19.9.0
Pillow>=6.2.0
psycopg2==2.8.3
psycopg2-binary==2.8.3
pycparser==2.19
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
text-unidecode==1.3
whitenoise==4.1.3
django_session_timeout == 0.0.4
lxml == 4.4.1
django-tables2==2.1.1
django-filter==2.2.0
tablib==0.14.0
django-bootstrap3==11.1.0
django-bootstrap4==1.0.1
django-crispy-forms==1.8.0
